add TerminalControlBar and TerminalGraph components
