+++ src/components/TerminalWindow.vue
@@ -18,18 +18,13 @@
       ref="terminalWindow"
       @click="focusInput"
     >
-      <div
-        class="terminal-header"
-        @mousedown="startDrag"
-        @mouseup="stopDrag"
-      >
-        <div class="terminal-controls">
-          <div class="control close" @click="$emit('close')"></div>
-          <div class="control minimize" @click="isMinimized = true"></div>
-          <div class="control maximize"></div>
-        </div>
-        <div class="terminal-title">{{ title }}</div>
-      </div>
+      <TerminalControlBar
+        :title="title"
+        @close="$emit('close')"
+        @minimize="isMinimized = true"
+        @startDrag="startDrag"
+        @stopDrag="stopDrag"
+      />
 
       <div class="terminal-content" @click="focusInput">
         <div class="terminal-output" ref="terminalOutput">
@@ -45,57 +40,21 @@
                 <template v-else-if="line.type === 'link'">
                   {{ line.prefix }}<a :href="line.url" class="terminal-link">{{ line.text }}</a>
                 </template>
-                <template v-else-if="line.type === 'graph-history'">
-                  <span class="git-graph-date">{{ line.date }}:</span>
-                  <span class="git-graph-additions">{{ line.additionBars }}</span>
-                  <span class="git-graph-deletions">{{ line.deletionBars }}</span>
-                  <span class="git-graph-stats">({{ line.additions }} additions, {{ line.deletions }} deletions)</span>
-                </template>
-                <template v-else-if="line.type === 'latest-commit'">
-                  <div class="git-latest-commit-container">
-                    <div class="git-graph-title">{{ line.title }}</div>
-                    <div class="latest-commit-content">
-                      <div class="latest-commit-line">
-                        <span class="latest-commit-hash">[{{ line.hash }}]</span>
-                        <a v-if="line.url" :href="line.url" class="terminal-link">View on GitHub</a>
-                      </div>
-                      <div class="latest-commit-message">{{ line.message }}</div>
-                    </div>
-                  </div>
-                </template>
+                <terminal-graph-output
+                  v-else-if="line.type === 'graph-history' || line.type === 'latest-commit'"
+                  :line="line"
+                />
               </div>
 
-              <div v-if="item.graphData && item.graphData.isVisible" class="git-graph-container">
-                <div class="git-graph-title">{{ item.graphData.title }}</div>
+              <terminal-graph-output
+                v-if="item.graphData && item.graphData.isVisible"
+                :graph-data="item.graphData"
+              />
 
-                <div v-if="item.graphData.noData" class="git-graph-no-data">
-                  <div class="git-graph-no-data-icon">📊</div>
-                  <div class="git-graph-no-data-message">No code frequency data available</div>
-                  <div class="git-graph-no-data-explanation">{{ item.graphData.note }}</div>
-                </div>
-
-                <template v-else>
-                  <div v-for="(week, weekIndex) in item.graphData.weeks" :key="weekIndex" class="git-graph-row">
-                    <span class="git-graph-date">{{ week.date }}:</span>
-                    <span class="git-graph-additions">{{ '+'.repeat(week.additionBars) }}</span>
-                    <span class="git-graph-deletions">{{ '-'.repeat(week.deletionBars) }}</span>
-                    <span class="git-graph-stats">({{ week.additions }} additions, {{ week.deletions }} deletions)</span>
-                  </div>
-                </template>
-
-                <div v-if="!item.graphData.noData" class="git-graph-note">{{ item.graphData.note }}</div>
-              </div>
-
-              <div v-if="item.commitData && item.commitData.isVisible" class="git-latest-commit-container">
-                <div class="git-graph-title">{{ item.commitData.title }}</div>
-                <div class="latest-commit-content">
-                  <div class="latest-commit-line">
-                    <span class="latest-commit-hash">[{{ item.commitData.hash }}]</span>
-                    <a v-if="item.commitData.url" :href="item.commitData.url" class="terminal-link">View on GitHub</a>
-                  </div>
-                  <div class="latest-commit-message">{{ item.commitData.message }}</div>
-                </div>
-              </div>
+              <terminal-graph-output
+                v-if="item.commitData && item.commitData.isVisible"
+                :commit-data="item.commitData"
+              />
 
               <div v-if="item.isLoading" class="loading-container">
                 <div class="progress-bar-container">
@@ -132,7 +91,8 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { getLatestCommitMessage, getCodeFrequency } from '../utils/gitInfo.js';
 import { navItems, commandHistoryStore, nextCommandIdStore } from '../stores/ui.js';
 import { useStore } from '@nanostores/vue';
-// We don't need to import ClientOnly here anymore.
+import TerminalControlBar from './TerminalControlBar.vue';
+import TerminalGraphOutput from './TerminalGraphOutput.vue';
 
 library.add(faTerminal);
 
@@ -140,7 +100,8 @@ export default {
   name: 'TerminalWindow',
   components: {
     FontAwesomeIcon,
-    // ClientOnly is no longer a component here.
+    TerminalControlBar,
+    TerminalGraphOutput,
   },
   props: {
     title: {
@@ -395,12 +356,14 @@ export default {
             .then(frequencyData => {
               if (historyIndex !== -1) {
                 const updatedHistory = [...commandHistory.value];
-                const graphData = processCodeFrequencyData(frequencyData);
+                // Create a temporary instance to use the method
+                const graphOutputRef = TerminalGraphOutput.methods.processCodeFrequencyData(frequencyData);
+
                 updatedHistory[historyIndex] = {
                   ...updatedHistory[historyIndex],
                   textOutput: [...updatedHistory[historyIndex].textOutput, 'Code Frequency (additions/deletions over time):', ''],
                   isLoading: false,
-                  graphData: graphData
+                  graphData: graphOutputRef
                 };
                 commandHistoryStore.set(updatedHistory);
                 setTimeout(() => { if (terminalOutput.value) terminalOutput.value.scrollTop = terminalOutput.value.scrollHeight; }, 0);
@@ -429,45 +392,7 @@ export default {
       }
     };
 
-    const processCodeFrequencyData = (frequencyData) => {
-      if (frequencyData && frequencyData.computing) {
-        return { title: 'Code Frequency Data', weeks: [], note: frequencyData.message, isVisible: true, noData: true };
-      }
-      if (frequencyData && frequencyData.error) {
-        return { title: 'Code Frequency Data', weeks: [], note: frequencyData.message, isVisible: true, noData: true };
-      }
-      if (!Array.isArray(frequencyData)) {
-        return { title: 'Code Frequency Data', weeks: [], note: 'Invalid data format received', isVisible: true, noData: true };
-      }
-      if (frequencyData.length === 0) {
-        return { title: 'Code Frequency Data', weeks: [], note: 'No code frequency data available. This could be because the repository is new, private, or the GitHub API has not calculated the statistics yet.', isVisible: true, noData: true };
-      }
-
-      const recentData = frequencyData.slice(-10);
-      let maxAddition = 0;
-      let maxDeletion = 0;
-      recentData.forEach(week => {
-        maxAddition = Math.max(maxAddition, week[1]);
-        maxDeletion = Math.max(maxDeletion, Math.abs(week[2]));
-      });
-      const maxValue = Math.max(maxAddition, maxDeletion);
-      const graphHeight = 10;
-
-      return {
-        title: 'Additions (+) / Deletions (-) - Last 10 weeks',
-        weeks: recentData.map(week => {
-          const date = new Date(week[0] * 1000).toISOString().split('T')[0];
-          const additions = week[1];
-          const deletions = Math.abs(week[2]);
-          const additionBars = Math.round((additions / maxValue) * graphHeight);
-          const deletionBars = Math.round((deletions / maxValue) * graphHeight);
-          return { date, additions, deletions, additionBars, deletionBars };
-        }),
-        note: 'Note: Graph is scaled to fit the terminal window',
-        isVisible: true,
-        noData: false
-      };
-    };
+    // processCodeFrequencyData function moved to TerminalGraphOutput.vue
 
     const handleDocumentClick = (event) => {};
 
@@ -495,7 +420,6 @@ export default {
       terminalWindow,
       terminalInput,
       terminalOutput,
-      processCodeFrequencyData,
       focusInput,
       isLoading,
       loadingProgress,
@@ -563,48 +487,7 @@ export default {
   z-index: 1000;
 }
 
-/* Terminal header with controls */
-.terminal-header {
-  height: 28px;
-  background-color: #3c3c3c;
-  display: flex;
-  align-items: center;
-  padding: 0 10px;
-  cursor: move;
-  user-select: none;
-}
-
-.terminal-controls {
-  display: flex;
-  gap: 8px;
-}
-
-.control {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  cursor: pointer;
-}
-
-.close {
-  background-color: #ff5f56;
-}
-
-.minimize {
-  background-color: #ffbd2e;
-}
-
-.maximize {
-  background-color: #27c93f;
-}
-
-.terminal-title {
-  flex-grow: 1;
-  text-align: center;
-  color: #cccccc;
-  font-size: 14px;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
+/* Terminal header styles moved to TerminalControlBar.vue */
 
 /* Terminal content area */
 .terminal-content {
@@ -678,13 +561,7 @@ export default {
   color: #333;
 }
 
-.terminal-window.theme-light .terminal-header {
-  background-color: #d8d8d8;
-}
-
-.terminal-window.theme-light .terminal-title {
-  color: #333;
-}
+/* Terminal header theme styles moved to TerminalControlBar.vue */
 
 .terminal-window.theme-light .terminal-content {
   color: #333;
@@ -726,160 +603,7 @@ export default {
   background: #aaa;
 }
 
-/* Git graph styling */
-.git-graph-additions {
-  color: #27c93f; /* Green color for additions */
-  font-weight: bold;
-  letter-spacing: 1px;
-}
-
-.git-graph-deletions {
-  color: #ff5f56; /* Red color for deletions */
-  font-weight: bold;
-  letter-spacing: 1px;
-}
-
-.git-graph-date {
-  color: #ffbd2e; /* Yellow color for dates */
-  font-weight: bold;
-}
-
-.git-graph-stats {
-  color: #cccccc; /* Light gray for stats */
-  font-style: italic;
-  font-size: 0.9em;
-}
-
-.git-graph-title {
-  color: #ffffff; /* White color for title */
-  font-weight: bold;
-  font-size: 1.1em;
-  text-decoration: underline;
-  margin-bottom: 10px;
-}
-
-.git-graph-note {
-  color: #aaaaaa; /* Light gray for note */
-  font-style: italic;
-  font-size: 0.85em;
-}
-
-/* Theme-specific git graph styling */
-.theme-light .git-graph-additions {
-  color: #27a83f;
-}
-
-.theme-light .git-graph-deletions {
-  color: #e74c3c;
-}
-
-.theme-light .git-graph-date {
-  color: #f39c12;
-}
-
-.theme-light .git-graph-stats {
-  color: #666666;
-}
-
-.theme-light .git-graph-title {
-  color: #333333;
-}
-
-.theme-light .git-graph-note {
-  color: #777777;
-}
-
-/* Git graph container styling */
-.git-graph-container {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  padding: 5px;
-  border: 1px solid #444;
-  border-radius: 4px;
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
-.git-graph-row {
-  display: flex;
-  align-items: center;
-  margin: 2px 0;
-  white-space: nowrap;
-}
-
-/* No data message styling */
-.git-graph-no-data {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 20px 10px;
-  text-align: center;
-}
-
-.git-graph-no-data-icon {
-  font-size: 24px;
-  margin-bottom: 10px;
-}
-
-.git-graph-no-data-message {
-  font-size: 14px;
-  color: #cccccc;
-  font-style: italic;
-  margin-bottom: 8px;
-}
-
-.git-graph-no-data-explanation {
-  font-size: 12px;
-  color: #aaaaaa;
-  font-style: italic;
-  max-width: 80%;
-  text-align: center;
-  margin-top: 8px;
-}
-
-.theme-light .git-graph-container {
-  border-color: #ccc;
-  background-color: rgba(0, 0, 0, 0.05);
-}
-
-.theme-light .git-graph-no-data-message {
-  color: #666666;
-}
-
-.theme-light .git-graph-no-data-explanation {
-  color: #888888;
-}
-
-/* Git latest commit container styling - similar to git graph container */
-.git-latest-commit-container {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  padding: 5px;
-  border: 1px solid #444;
-  border-radius: 4px;
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
-.theme-light .git-latest-commit-container {
-  border-color: #ccc;
-  background-color: rgba(0, 0, 0, 0.05);
-}
-
-.latest-commit-line {
-  margin: 5px 0;
-}
-
-.latest-commit-hash {
-  color: #ffbd2e; /* Yellow color for hash, same as dates in graph */
-  font-weight: bold;
-  margin-right: 10px;
-}
-
-.latest-commit-message {
-  margin-top: 8px;
-  white-space: pre-wrap;
-  word-break: break-all;
-}
+/* Git graph styling moved to TerminalGraphOutput.vue */
 
 /* Responsive adjustments */
 @media (max-width: 768px) {
